Resolved #2109 This new script should update asset panels, based on changes reported or new information at intake.
This script will either add a new ACCT panel to MAXIS or update an existing ACCT panel. It can be ran multiple times to either add/update multiple ACCT panels in any combination of add/update. It will then case note all actions upon completion of the script.

This is part of the ongoing Project Bowie.