Resolved #2412 Create- Notes- Paris Match
This is a script that case notes a PARIS match action. This can be used to case note if the match was resolved or if a notice was sent to the client. Worker also indicates the month of the match, Household member, State where the match occurred, the matched program, and the resolution code used.