Resolves #1142 -- Creating Inverse Krabappel
BLIP: Adding the "Inverse Krabappel" script file. This script will grab case data from a training case and/or production cases, and will put the data into an Excel spreadsheet. For the script to run all the way through, you need to use a Krabappel template. The script is programmed to run with the data validation on the cells. INSTRUCTIONS TO COME!!!