ADMIN - FSET sanctions: Updates for a single master list and case notes updates
• Removed agency informed of sanction field from the dialog
• Added dialog to select which excel row to start (great for processing
sanction cases as they come in)
• Created constants in the excel columns for easy updates
• Updated coding for PRIV cases. Was putting in OUT OF COUNTY indicator
as well.
• Added 30/11 cases as mandatory sanction, so if referrals were made and
letters were sent, the cases will be identified as "YES" to sanction.
• Added info to case note that client does not appear to meet Good
Cause. I based this off the information provided in the MAXIS mentor
meeting about eligibility workers needing to determine good cause.
• Added information to the case note about ABAWD exemptions. If the case
is identified in the notes as 'Possible homeless exemption' then the
case note will include that they may be eligible for this exemption.
Otherwise the case notes will reflect that "Client does not appear to
meet any ABAWD exemptions."