#331 TASK BASED DAIL CAPTURE: Changed DAIL navigation
In testing there were issues with a number of messages that although in the DAIL count on the panel, were not visable by navigating with PF7/PF8. Navigation needed to revert back to the 'T' process, which is what is in other DAIL based BULK scripts.

Also added back in Excel output for testing, even though those messages are not being deleted any longer.