#381 NOTES OVERPAYMENT - Requesting Claims Adjustment
Claims: Underpayments, Adjustments, and Closing
Can I get a BZ script for claim adjustment. This is a script that we have to case note when a claim was already entered and we need accounting to make a correction on the claim.

Requested Claim Adjustment*
Claim number:
Period:
Original Amount:
Correct Amount:
Reason for correction: