#140 NOTES - AVS: Added additional AVS status of N/A
This will be used for instances where members do not need to submit an AVS request as they are PRIV, not applying, not deeming, etc.
Added conditional logic to ensure that the users must enter the reason for N/A and the Request Type also must match N/A.

Thank you @C-Love for creating awareness of this process!