#473 Updated MMIS CASE NOTE to DHS COMMENTS on ADHS panel
Project sponsor, who is going to be primary user, does not have MMIS access to the C (or health care) realm. For this reason DHS Comments will be added.
Created an array since there could already be notes in the DHS Comments section.
Uncommented out MAXIS case notes for testing.