#558 ADMIN - TIKL FROM LIST: added pass through variable for custom function
manual entry TIKL date was not carrying through to the TIKL function. There was a pass through variable for the case number and the TIKL text, but not the date.  Corrected that.
Also updated the TIKL_text and TIKL_date to variables not used in the create_TIKL Custom functions