#814 DAIL - NEW HIRE: FS PIC issue
FS_active was not determined, and hired month was not defined, causing an issue in updating the PIC for SNAP. Removed condition, and now will update for all programs.
Removed unneeded coding around bringing DAIL to the top of the DAIL, and still on DAIL as this will ease the transition to a single script.
Added write_value_and_transmits as this is will ease the transition to a single script.
Updated supports to exit the PIC and get back to self.