#814 DAIL - NEW HIRE NDNH: FS PIC Issue
FS_active was not determined, and hired month was not defined, causing an issue in updating the PIC for SNAP. Removed condition, and now will update for all programs.
Added write_value_and_transmits as this is will ease the transition to a single script.
Updated coding to save and leave the PIC and get back to self.
Updated closing message for NEW HIRE to inform user if not carrying the panel information to the next footer month.