#927 NOTES RETURNED MAIL bugs
Updated Line 199 "Verification(s) pending:" - If we are requesting verifications this should say 'Verification (s) requested'.

Updated to inhibit the selection of  an address that is blank.
other_notes - update the tab order.
Adjusted TIKL to make a complete sentence.
Removed if statement specific to cash and snap around the PACT panel dialog and and added other notes.
Line 397 and others: Add periods to case/note when using wirte_variable_in_case_note