#983 DEU - BULK MATCH CLEARED: enhanced comments in Excel output
The comments aren't necessarily helpful to end users, and a bug was found when a match wasn't pending. Updated code to have a pending_match_found = False default and will add comments for the user to review the case.

Added msgboxes in since this may need to be reviewed again.