#1449 Notes HC Transition- Aligned case/note and spec/memo with previous setup
Mets_case_number will now be mets_or_pr_number (both METS# and PR#). To align case/note and spec/memo with previous structure, if the field is blank regardless of whether METS# or PR# was selected from the drop down, it will provide the same results as mets_case_number = "" in the previous version.