NOTES - INTERVIEW: Making MS Word visible for users
DWP staff have been having issues using the interview script, specifically SH. There is a known access issue with DWP staff, and several have tickets open to fix something in MS Excel. Wondering if this is related since it's in the O365 suite, so just stabbing the dark at this point.